+++ .gitignore
@@ -85,7 +85,6 @@ instance/
 
 # Sphinx documentation
 sphinx-docs/build/
-sphinx-docs/sphinx-build/
 sphinx-docs/source/md/
 sphinx-docs/source/docs/CHANGELOG.md
 sphinx-docs/source/docs/README.md
